contracts-stylus: core: Split `core-settlement` into three contracts
