+++ contracts-stylus/src/contracts/core/mod.rs
@@ -14,10 +14,15 @@ use crate::{
 
 #[cfg(any(feature = "darkpool-core", feature = "darkpool-test-contract"))]
 pub mod core_helpers;
-#[cfg(feature = "core-settlement")]
-pub mod core_settlement;
 #[cfg(any(feature = "core-wallet-ops", feature = "darkpool-test-contract"))]
 pub mod core_wallet_ops;
+
+#[cfg(feature = "core-atomic-match-settle")]
+pub mod core_atomic_match_settle;
+#[cfg(feature = "core-malleable-match-settle")]
+pub mod core_malleable_match_settle;
+#[cfg(feature = "core-match-settle")]
+pub mod core_match_settle;
 
 /// A trait that allows for storage access to the standard storage layout for
 /// core contracts
