+++ contracts-stylus/src/utils/constants.rs
@@ -75,13 +75,21 @@ pub const INVALID_TRANSACTION_VALUE_ERROR_MESSAGE: &[u8] = b"transaction value s
 
 /// The revert message when order settlement indices are different
 /// between a VALID COMMITMENTS & VALID MATCH SETTLE statement
-#[cfg(feature = "core-settlement")]
+#[cfg(any(
+    feature = "core-match-settle",
+    feature = "core-atomic-match-settle",
+    feature = "core-malleable-match-settle"
+))]
 pub const INVALID_ORDER_SETTLEMENT_INDICES_ERROR_MESSAGE: &[u8] =
     b"invalid order settlement indices";
 
 /// The revert message when the protocol fee is incorrect in a
 /// VALID MATCH SETTLE statement
-#[cfg(feature = "core-settlement")]
+#[cfg(any(
+    feature = "core-match-settle",
+    feature = "core-atomic-match-settle",
+    feature = "core-malleable-match-settle"
+))]
 pub const INVALID_PROTOCOL_FEE_ERROR_MESSAGE: &[u8] = b"invalid protocol fee";
 
 /// The revert message when the protocol public encryption key is
